Database and Class Setups
I have created the class files required (1 extra that wasn't in the spec
but was in the access file).

The columns "Priority" in Tasks and "Status" in Projects are in the
database as ints, I have done this with the plan for those values in the
classes to be put in as Enums. It will make our lives easier later on
(including mine with storing stuff in the database).