Changed database access to linq
I have changed all the database access to be handled through LINQ, this
makes it a 2-3 line process per method instead of 10-15. I have also
made it so all Select, Add, Update and Delete methods are generalised.
This means no matter what table you are accessing, it is the same
method. The one exception is getting one specific IterationTask or all
Tasks for an Iteration (same table, just couldnt make them generic
methods).

If you have used the previous database calls (only the developers page
as of this commit and i have updated those) then it is a simple task to
change them to the new ones.

See you all Tuesday!