Use Data Access Object design for handling data.
The DBInterface was bloated so I've stripped out handling of Developer from it to use the DAO design pattern. So rather than use switches and if statements a new DAO interface and concrete class is created for each table. Then the DAOFactory is used to get the appropriate DAO class. At present only SQL Server DAO's are required but this may need to be modified when we want to do some testing.