Use VARCHAR, Not Null and default value in Table design.
Rather than use TEXT which cannot be sorted replaced it with VARCHAR data type, using commonly accepted lengths.
Not Null constraints added on fields that must be entered for new and updated data.
The Created field in Clients uses current date by default as it's just one less thing to worry about in the application.