Disable Save and Cancel buttons.
The Save button is disabled when required fields (Given Name, Family Name, Email, Contact Number) are empty. As soon as they all have a value the Save button is enabled.
The Cancel button is disabled when adding a new developer and all required fields are empty. As soon as text is entered the Cancel button is enabled.