Added graph UI and test data
Just some testing on the graph UI, seems to work quite well, will have
to be refined once all UI's are in place and able to generate the
report.

Haven't added the print function yet but it needs to be able to print to
a printer according to the spec, will look at that more tomorrow.

Made a small change to the DBInterface for getting all tasks for an
iteration.