Adding and Editing of Developers added to own form.
Moved all the Text Fields for Adding and Editing Developers to a new form to prepare for new layout of listing developer iterations and tasks.
The Cancel button on new form always visible, Delete is only available for editing developers and has not been implemented. Save is enabled when all required fields have a value.
If the DeveloperForm is closed (hidden) also close (hide) EditDeveloperForm.
When EditDeveloperForm is closed (hidden) update the DeveloperForm so any changes are reflected in the ListView.
The DeveloperForm also uses ItemActivate on OneClick instead of SelectedIndexChanged event.