Add list of Iterations to EditDeveloperForm in a DataGridView.
Created a new View to get the list of iterations for developers and added it to the Linq to SQL Generator.
When editing a developer the DataGridView is populated with the iterations, on adding  developer it is hidden.
The first column is to print a report, but at present only shows a MessageBox.
Added some more tasks and developer iteration tasks to better test the system.