Add developer to task form complete
Changed the way the datacontext is created. creating a new instance was
creating concurrency issues and not allowing creation of some items. If
you are going to delete an item from the database, you need to first
retrieve the exact copy of that item with the Get...() methods otherwise
the database doesnt think it exists. Message me for more info if
confused.