Refactored Developer Iteration Report to use methods rather than repeated code in one massive method.
A lot of the variables had to be made member variables to avoid unnecessary parsing of values from one method to the next.
New methods are self contained and named clearly.
Could be further refactored as DrawString is called a number of times when printing headers, paragraphs and paragraphs with headers.