Text is wrapped onto next page rather than it going to the edge of the page and being chopped off.
Bit tedious to get this working but all fields except for Completed checkbox (which should never extend entire line) doew word wrapping.
There is a limitation that only one page is printed, so any content that extends a page is lost. If anyone thinks I need to address this let me know and I'll investigate further.