Removed label and added Close button.
The label wasn't giving any information that is useful to the user, they have no idea what the iteration ID is.
The Close button closes the form. This does show a leak in how this form is loaded (ie there may be multiple copies) but I'll leave that for now.