Inactive tasks and iterations are removed from Developer views. Active column removed from Task section of Developers.
When getting tasks for iteration exclude active = true records.
When getting iterations exclude active = true records.
Added a new inactive task and made iteration 3 active to check they are not visible to developer Carmen Grantham.
The Active column was removed from the Tasks section of Edit Developer Form as user does not need to see this.