Delete iterations, improved itTask editing
Delete button has been added to the iterations form. When one of the
iterations is selected and delete button is pressed, the user will
confirm deletion. On OK the iteration will be deleted.
Editing which tasks can be assigned to an iteration now takes into
account whether or not a task has been completed with "(Completed)"
being next to the task name and also not showing up at all if not
assigned to an iteration.